requirements.txt
celery==4.4.5
coverage==4.5.4
cryptography==2.7
Django==2.2
django-environ==0.4.5
django-extensions==2.1.6
django-json-widget==0.2.0
GitPython~=3.1
gunicorn==19.9.0
ipdb
lizard==1.15.6
numpy==1.16.2
pandas==0.24.2
psycopg2-binary==2.8.3
python-dateutil==2.8.0
PyJWT==1.7.1
pytz==2019.1
redis==3.2.1
requests==2.20.1
whitenoise==4.1.4
flower==0.9.3  # celery monitoring
sentry-sdk==0.14.4
django-structlog==1.5.2
django-mptt==0.11.0
django-anymail~=7.1.0
djangorestframework==3.11.1
drf-nested-routers==0.91